requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
click==8.2.0
colorama==0.4.6
contourpy==1.3.1
crispy==0.8.0
crispy-bootstrap5==2025.4
cycler==0.12.1
Django==5.2.1
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
drf-yasg==1.21.10
fabio==2024.9.0
Faker==37.1.0
fonttools==4.57.0
form==0.0.1
greenlet==3.1.1
gunicorn==23.0.0
h5py==3.13.0
hdf5plugin==5.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==5.3.2
matplotlib==3.10.1
mypy_extensions==1.1.0
numpy==2.2.4
packaging==25.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
silx==2.2.2
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
xraydb==4.5.6
